rename "gits" to "meitrex"
