+++ settings.gradle
@@ -1,7 +1,7 @@
-rootProject.name = 'gits-common-test'
+rootProject.name = 'meitrex-common-test'
 
 sourceControl {
     gitRepository(uri('https://github.com/MEITREX/common')) {
-        producesModule('de.unistuttgart.iste.gits:gits-common')
+        producesModule('de.unistuttgart.iste.meitrex:meitrex-common')
     }
 }
