+++ src/main/java/de/unistuttgart/iste/meitrex/common/testutil/GraphQlTesterParameterResolver.java
@@ -1,6 +1,6 @@
-package de.unistuttgart.iste.gits.common.testutil;
+package de.unistuttgart.iste.meitrex.common.testutil;
 
-import de.unistuttgart.iste.gits.common.user_handling.LoggedInUser;
+import de.unistuttgart.iste.meitrex.common.user_handling.LoggedInUser;
 import lombok.SneakyThrows;
 import org.junit.jupiter.api.extension.*;
 import org.springframework.graphql.test.tester.GraphQlTester;
