+++ src/main/java/de/unistuttgart/iste/meitrex/common/testutil/InjectCurrentUserHeader.java
@@ -1,6 +1,6 @@
-package de.unistuttgart.iste.gits.common.testutil;
+package de.unistuttgart.iste.meitrex.common.testutil;
 
-import de.unistuttgart.iste.gits.common.user_handling.LoggedInUser;
+import de.unistuttgart.iste.meitrex.common.user_handling.LoggedInUser;
 
 import java.lang.annotation.*;
 import java.util.UUID;
